feat(cli): add `wmux daemon start|stop|status` (#1001)
Per the maintainer's scope note in discussion #1001: exactly these three
verbs, no first-run wizard, no service-manager integration.

Extracts the spawn/readiness chain out of src/main/daemon/launcher.ts into
src/shared/daemon/daemonLauncherCore.ts (zero import of electron), behind a
DaemonLauncherDeps seam. Electron main binds it with app.getAppPath() /
app.getVersion() / the native recovery dialog, unchanged in behavior. The CLI
binds it with script resolution relative to its own bundle location, its own
package.json version, and an always-refuse stale-PID policy (a headless
caller has no window for a dialog and no keyboard to ask).

status = daemon.ping, stop = daemon.shutdown — both pre-existing daemon RPCs,
no new termination path.
